[Boundary interfaces] Add a package file to hide boundary interfaces
Android publishes all public APIs that are not marked with the @hide
annotation.
Boundary interfaces are not supposed to be public to apps. In this CL we
add a file with the sole purpose of hiding the boundary interface
package from apps.

Bug: 788177
Change-Id: I58c9c7a8445c1a3fdfda3164735f5e2c755dfc44
Reviewed-on: https://chromium-review.googlesource.com/918861
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Gustav Sennton <gsennton@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#536719}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0293525c239e35235b4ea8375d84956b4185efad